fix(get): address review — clap-required paths, shared normalize_path, drop Option
Resolves Keats' review on #203:
- restore #[clap(required_unless_present = "glob")] on `dvs get` paths and drop
  the runtime bail!; keep the "restore everything -> --glob '**/*'" hint in the
  command help so the guidance isn't lost
- move normalize_path out of files::status into paths (shared by globbing + status)
- drop Option<Vec<PathBuf>> dir_filters in resolve_paths_for_get; use a plain Vec
  plus a had_explicit_paths bool (preserves "paths given but all escaped root
  -> matches nothing")
- fix the misleading comment: cwd scoping is applied later via cwd_relative
